requirements.txt
alembic>=1.16.5
sqlalchemy>=2.0.43
python-dotenv>=1.1.1
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
azure-storage-blob
aiohttp

# worker
pika
